backend/requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
psycopg2-binary==2.9.3
Faker
django-cors-headers
django-extensions==3.1.0
stripe==2.76.0
